+++ guillotina_rediscache/tests/test_cache.py
@@ -11,7 +11,7 @@ from guillotina_rediscache.interfaces import IRedisChannelUtility
 import asyncio
 
 
-async def test_cache_set(redis, dummy_guillotina, loop):
+async def test_cache_set(redis_container, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
     trns.added = trns.deleted = {}
@@ -20,7 +20,8 @@ async def test_cache_set(redis, dummy_guillotina, loop):
 
     await rcache.set('bar', oid='foo')
     # make sure it is in redis
-    assert serialize.loads(await rcache._redis.get(CACHE_PREFIX + 'foo')) == "bar"
+    assert serialize.loads(
+        await rcache._redis.get(CACHE_PREFIX + 'foo')) == "bar"
     # but also in memory
     assert rcache._memory_cache.get('foo') == 'bar'
     # and api matches..
@@ -29,7 +30,7 @@ async def test_cache_set(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
 
 
-async def test_cache_delete(redis, dummy_guillotina, loop):
+async def test_cache_delete(redis_container, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
     trns.added = trns.deleted = {}
@@ -38,7 +39,8 @@ async def test_cache_delete(redis, dummy_guillotina, loop):
 
     await rcache.set('bar', oid='foo')
     # make sure it is in redis
-    assert serialize.loads(await rcache._redis.get(CACHE_PREFIX + 'foo')) == "bar"
+    assert serialize.loads(
+        await rcache._redis.get(CACHE_PREFIX + 'foo')) == "bar"
     assert rcache._memory_cache.get('foo') == 'bar'
     assert await rcache.get(oid='foo') == 'bar'
 
@@ -49,7 +51,7 @@ async def test_cache_delete(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
 
 
-async def test_cache_clear(redis, dummy_guillotina, loop):
+async def test_cache_clear(redis_container, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
     trns.added = trns.deleted = {}
@@ -58,7 +60,8 @@ async def test_cache_clear(redis, dummy_guillotina, loop):
 
     await rcache.set('bar', oid='foo')
     # make sure it is in redis
-    assert serialize.loads(await rcache._redis.get(CACHE_PREFIX + 'foo')) == "bar"
+    assert serialize.loads(
+        await rcache._redis.get(CACHE_PREFIX + 'foo')) == "bar"
     assert rcache._memory_cache.get('foo') == 'bar'
     assert await rcache.get(oid='foo') == 'bar'
 
@@ -68,7 +71,7 @@ async def test_cache_clear(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
 
 
-async def test_invalidate_object(redis, dummy_guillotina, loop):
+async def test_invalidate_object(redis_container, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
     trns.added = trns.deleted = {}
@@ -78,7 +81,8 @@ async def test_invalidate_object(redis, dummy_guillotina, loop):
     await rcache.clear()
 
     await rcache.set('foobar', oid=content._p_oid)
-    assert serialize.loads(await rcache._redis.get(CACHE_PREFIX + content._p_oid)) == "foobar"
+    assert serialize.loads(
+        await rcache._redis.get(CACHE_PREFIX + content._p_oid)) == "foobar"
     assert rcache._memory_cache.get(content._p_oid) == 'foobar'
     assert await rcache.get(oid=content._p_oid) == 'foobar'
 
@@ -88,7 +92,7 @@ async def test_invalidate_object(redis, dummy_guillotina, loop):
     await cache.close_redis_pool()
 
 
-async def test_subscriber_invalidates(redis, dummy_guillotina, loop):
+async def test_subscriber_invalidates(redis_container, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
     trns.added = trns.deleted = {}
@@ -98,23 +102,26 @@ async def test_subscriber_invalidates(redis, dummy_guillotina, loop):
     await rcache.clear()
 
     await rcache.set('foobar', oid=content._p_oid)
-    assert serialize.loads(await rcache._redis.get(CACHE_PREFIX + content._p_oid)) == "foobar"
+    assert serialize.loads(
+        await rcache._redis.get(CACHE_PREFIX + content._p_oid)) == "foobar"
     assert rcache._memory_cache.get(content._p_oid) == 'foobar'
     assert await rcache.get(oid=content._p_oid) == 'foobar'
 
     assert content._p_oid in rcache._memory_cache
 
-    await rcache._redis.publish(app_settings['redis']['updates_channel'], serialize.dumps({
-        'tid': 32423,
-        'keys': [content._p_oid]
-    }))
+    await rcache._redis.publish(
+        app_settings['redis']['updates_channel'], serialize.dumps({
+            'tid': 32423,
+            'keys': [content._p_oid]
+        }))
     await asyncio.sleep(1)  # should be enough for pub/sub to finish
     assert content._p_oid not in rcache._memory_cache
 
     await cache.close_redis_pool()
 
 
-async def test_subscriber_ignores_trsn_on_invalidate(redis, dummy_guillotina, loop):
+async def test_subscriber_ignores_trsn_on_invalidate(
+        redis_container, dummy_guillotina, loop):
     await cache.close_redis_pool()
     trns = mocks.MockTransaction(mocks.MockTransactionManager())
     trns.added = trns.deleted = {}
@@ -124,7 +131,8 @@ async def test_subscriber_ignores_trsn_on_invalidate(redis, dummy_guillotina, lo
     await rcache.clear()
 
     await rcache.set('foobar', oid=content._p_oid)
-    assert serialize.loads(await rcache._redis.get(CACHE_PREFIX + content._p_oid)) == "foobar"
+    assert serialize.loads(
+        await rcache._redis.get(CACHE_PREFIX + content._p_oid)) == "foobar"
     assert rcache._memory_cache.get(content._p_oid) == 'foobar'
     assert await rcache.get(oid=content._p_oid) == 'foobar'
 
@@ -133,10 +141,11 @@ async def test_subscriber_ignores_trsn_on_invalidate(redis, dummy_guillotina, lo
     utility = getUtility(IRedisChannelUtility)
     utility.ignore_tid(5555)
 
-    await rcache._redis.publish(app_settings['redis']['updates_channel'], serialize.dumps({
-        'tid': 5555,
-        'keys': [content._p_oid]
-    }))
+    await rcache._redis.publish(
+        app_settings['redis']['updates_channel'], serialize.dumps({
+            'tid': 5555,
+            'keys': [content._p_oid]
+        }))
     await asyncio.sleep(1)  # should be enough for pub/sub to finish
     # should still be there because we set to ignore this tid
     assert content._p_oid in rcache._memory_cache